refactor(money-formatters): simplify test and move files
