+++ src/shared/ui/utils/money-formatters/money-formatters.ts
